Test scripts that showed phase detection ability
Ran algorithm with the first part of the song ranging in small steps
from 59-61 seconds, with a Hanning window and with just a rectangular
window. Because of the different lengths, in most lengths, 130 (the
actual BPM for the song) was not a frequency bin, so a linear
interpolation was then done (accounting for the modular nature of the
phase from -pi to pi) to get estimates for the phase of 130 BPM.  The
Hanning window had very little variance between different phases, and
the rectangular window had much more.
With a Hanning window, this was repeated using different delays, ranging
from 0 to 60 seconds. Adding the theoretical phase change from the delay
to the measured phase, for delays of up to somewhere between 34 and 36
seconds, the phases were within .01 seconds of the phase of no delay.
In conclusion, the Hanning window and this algorithm allow accurate beat
and phase calculation.